move ingredient filter into callbacks and components
